rename(auth): auth -> user
목적에 맞게 디렉토리명을 auth -> user로 변경했습니다.
추후 해당 경로에 Employee 및 Manager 등 회원 관련 기능이 모두 해당 경로로 이동 또는 구현될 예정입니다.
