--- src/main/java/org/triumers/kmsback/auth/command/Application/dto/PasswordDTO.java
@@ -1,13 +1,0 @@
-package org.triumers.kmsback.auth.command.Application.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@ToString
-public class PasswordDTO {
-    private String oldPassword;
-    private String newPassword;
-}
