--- src/main/java/org/triumers/kmsback/auth/command/Application/service/AuthService.java
@@ -1,18 +1,0 @@
-package org.triumers.kmsback.auth.command.Application.service;
-
-import org.triumers.kmsback.auth.command.Application.dto.AuthDTO;
-import org.triumers.kmsback.auth.command.Application.dto.PasswordDTO;
-import org.triumers.kmsback.auth.command.domain.aggregate.entity.Employee;
-import org.triumers.kmsback.common.exception.NotLoginException;
-import org.triumers.kmsback.common.exception.WrongInputTypeException;
-import org.triumers.kmsback.common.exception.WrongInputValueException;
-
-public interface AuthService {
-    void signup(AuthDTO authDTO);
-
-    void editPassword(PasswordDTO passwordDTO) throws WrongInputTypeException, WrongInputValueException, NotLoginException;
-
-    void editMyInfo(AuthDTO authDTO) throws WrongInputTypeException, NotLoginException;
-
-    Employee whoAmI() throws NotLoginException;
-}
