--- src/main/java/org/triumers/kmsback/auth/command/domain/aggregate/enums/UserRole.java
@@ -1,8 +1,0 @@
-package org.triumers.kmsback.auth.command.domain.aggregate.enums;
-
-public enum UserRole {
-    ROLE_ADMIN,
-    ROLE_HR_MANAGER,
-    ROLE_LEADER,
-    ROLE_NORMAL
-}
