--- src/main/java/org/triumers/kmsback/auth/command/domain/aggregate/enums/UserRoleConverter.java
@@ -1,16 +1,0 @@
-package org.triumers.kmsback.auth.command.domain.aggregate.enums;
-
-import jakarta.persistence.AttributeConverter;
-
-public class UserRoleConverter implements AttributeConverter<UserRole, String> {
-
-    @Override
-    public String convertToDatabaseColumn(UserRole userRole) {
-        return userRole.name();
-    }
-
-    @Override
-    public UserRole convertToEntityAttribute(String dbData) {
-        return UserRole.valueOf(dbData);
-    }
-}
