--- src/main/java/org/triumers/kmsback/auth/command/domain/aggregate/vo/CmdResponseMessageVO.java
@@ -1,12 +1,0 @@
-package org.triumers.kmsback.auth.command.domain.aggregate.vo;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@ToString
-public class CmdResponseMessageVO {
-    private String message;
-}
