--- src/main/java/org/triumers/kmsback/auth/command/domain/repository/EmployeeRepository.java
@@ -1,10 +1,0 @@
-package org.triumers.kmsback.auth.command.domain.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import org.triumers.kmsback.auth.command.domain.aggregate.entity.Employee;
-
-@Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
-    Employee findByEmail(String email);
-}
