+++ src/main/java/org/triumers/kmsback/employee/command/Application/service/CmdEmployeeService.java
@@ -2,11 +2,13 @@ package org.triumers.kmsback.employee.command.Application.service;
 
 import org.triumers.kmsback.employee.command.Application.dto.CmdEmployeeDTO;
 
+import java.util.List;
+
 public interface CmdEmployeeService {
 
     CmdEmployeeDTO findEmployeeById(int id);
 
     CmdEmployeeDTO findEmployeeByEmail(String email);
 
-    CmdEmployeeDTO findEmployeeByName(String name);
+    List<CmdEmployeeDTO> findEmployeeByName(String name);
 }
