+++ src/main/java/org/triumers/kmsback/user/command/domain/aggregate/enums/UserRole.java
@@ -1,0 +1,8 @@
+package org.triumers.kmsback.user.command.domain.aggregate.enums;
+
+public enum UserRole {
+    ROLE_ADMIN,
+    ROLE_HR_MANAGER,
+    ROLE_LEADER,
+    ROLE_NORMAL
+}
