+++ src/main/java/org/triumers/kmsback/user/command/domain/aggregate/enums/UserRoleConverter.java
@@ -1,0 +1,16 @@
+package org.triumers.kmsback.user.command.domain.aggregate.enums;
+
+import jakarta.persistence.AttributeConverter;
+
+public class UserRoleConverter implements AttributeConverter<UserRole, String> {
+
+    @Override
+    public String convertToDatabaseColumn(UserRole userRole) {
+        return userRole.name();
+    }
+
+    @Override
+    public UserRole convertToEntityAttribute(String dbData) {
+        return UserRole.valueOf(dbData);
+    }
+}
