+++ src/main/java/org/triumers/kmsback/user/command/domain/aggregate/vo/CmdResponseMessageVO.java
@@ -1,0 +1,12 @@
+package org.triumers.kmsback.user.command.domain.aggregate.vo;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+public class CmdResponseMessageVO {
+    private String message;
+}
