+++ src/main/java/org/triumers/kmsback/user/command/domain/repository/EmployeeRepository.java
@@ -1,0 +1,10 @@
+package org.triumers.kmsback.user.command.domain.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import org.triumers.kmsback.user.command.domain.aggregate.entity.Employee;
+
+@Repository
+public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+    Employee findByEmail(String email);
+}
