+++ src/main/java/org/triumers/kmsback/user/command/domain/service/CustomUserDetailsService.java
@@ -1,0 +1,34 @@
+package org.triumers.kmsback.user.command.domain.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+import org.triumers.kmsback.user.command.Application.dto.CustomUserDetails;
+import org.triumers.kmsback.user.command.domain.aggregate.entity.Employee;
+import org.triumers.kmsback.user.command.domain.repository.EmployeeRepository;
+
+@Service
+public class CustomUserDetailsService implements UserDetailsService {
+
+    private final EmployeeRepository employeeRepository;
+
+    @Autowired
+    public CustomUserDetailsService(EmployeeRepository employeeRepository) {
+        this.employeeRepository = employeeRepository;
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+
+        Employee userData = employeeRepository.findByEmail(email);
+
+        if (userData != null) {
+
+            return new CustomUserDetails(userData);
+        }
+
+        throw new UsernameNotFoundException("존재하지 않는 유저입니다.");
+    }
+}
